Support Java 16 for Graal 21.1.0
Implements support for Java 16 as Graal 21.1.0 is compatible with it,
see https://www.infoq.com/news/2021/05/graalvm-21-1-supports-jdk16/

The download URL is the same as with Java 11 so this change should make
it work. It can be downloaded manually at https://github.com/graalvm/graalvm-ce-builds/releases/download/vm-21.1.0/graalvm-ce-java16-windows-amd64-21.1.0.zip

In addition to these changes, this commit also refactors
`isGraalVersionGreaterThan19_3` to support checks against 21.1 and also
includes `Equal` in the name as the previous name is a misnomer as it's
actually checking for "greater or equal than" instead of just "greater
than".